+++ src/index.ts
@@ -6,11 +6,11 @@ enum ENV {
   "development" = "development",
   "local" = "local",
 }
-
+console.log("process.env.NODE_ENV:", process.env.NODE_ENV);
 const envPath = `.env.${ENV[process.env.NODE_ENV]}`;
 
 require("dotenv").config({ path: envPath });
-console.log("process.env", process.env);
+// console.log("process.env", process.env);
 
 let server: Server = null;