+++ src/interface/user.interface.ts
@@ -1,0 +1,8 @@
+//import { Role } from "../interface/role";
+
+export interface User {
+  _id: number;
+  name?: string;
+  //roles: Role[];
+  roles: string[];
+}