+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # changelog
 
 ### 1.0.0
+
 - add env,github #2
+- add auth jwt role , add bcryptjs cookie-parser jsonwebtoken @types/bcryptjs @types/jsonwebtoken @types/cookie-parser #3